e2e/pyproject.toml
[tool.poetry]
name = "e2e"
version = "0.1.0"
description = "End-to-end tests for immunization-fhir-api"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.8"


[tool.poetry.group.dev.dependencies]
lxml = "^5.1.0"
requests = "^2.31.0"
pyjwt = "^2.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
